+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Please add changes at the top. When releasing a version add a new header for tha
 
 ### master
 
+ - New vertex operation "Add rank attribute" make it possible for the user to sort an
+   attribute (of double type). This can be used to identify vertices which have the
+   highest of lowest values with respect to some attributes.
  - Improved error checking for JavaScript expressions. (In "derive" operations.)
  - Fixed project history performance issue.
  - More than one type of aggregation can be selected for each attribute in one operation.